chat entities moved out of folder
