+++ MangoAPI.BusinessLogic/ApiCommands/Users/RegisterCommandHandler.cs
@@ -5,7 +5,6 @@ using MangoAPI.Application.Interfaces;
 using MangoAPI.BusinessLogic.Responses;
 using MangoAPI.Domain.Constants;
 using MangoAPI.Domain.Entities;
-using MangoAPI.Domain.Entities.ChatEntities;
 using MangoAPI.Domain.Enums;
 using MangoAPI.Infrastructure.Database;
 using MediatR;
