+++ MangoAPI.UnitTests/Domain/ChatEntityTests/ChatEntityCreateTestSuccess.cs
@@ -1,5 +1,5 @@
 ﻿using FluentAssertions;
-using MangoAPI.Domain.Entities.ChatEntities;
+using MangoAPI.Domain.Entities;
 using MangoAPI.UnitTests.Helpers;
 using System;
 using Xunit;
