+++ MangoAPI.UnitTests/Domain/UserChatEntityTests/UserChatEntityTestSuccess.cs
@@ -1,5 +1,5 @@
 ﻿using FluentAssertions;
-using MangoAPI.Domain.Entities.ChatEntities;
+using MangoAPI.Domain.Entities;
 using MangoAPI.Domain.Enums;
 using System;
 using Xunit;
