+++ MangoAPI.UnitTests/Domain/UserChatEntityTests/UserChatEntityTestsShouldThrow.cs
@@ -1,6 +1,6 @@
 ﻿using FluentAssertions;
 using FluentValidation;
-using MangoAPI.Domain.Entities.ChatEntities;
+using MangoAPI.Domain.Entities;
 using MangoAPI.Domain.Enums;
 using System;
 using Xunit;
